ContainerRegistry: Make digest(of:Data) an ImageReference.Digest constructor
