+++ Sources/containertool/Extensions/RegistryClient+Layers.swift
@@ -52,7 +52,7 @@ extension RegistryClient {
         mediaType: String = "application/vnd.oci.image.layer.v1.tar+gzip"
     ) async throws -> ImageLayer {
         // The diffID is the hash of the unzipped layer tarball
-        let diffID = digest(of: contents)
+        let diffID = ImageReference.Digest(of: contents)
         // The layer blob is the gzipped tarball;  the descriptor is the hash of this gzipped blob
         let blob = Data(gzip(contents))
         let descriptor = try await putBlob(repository: repository, mediaType: mediaType, data: blob)
